logging: add option to use android logger for process logs (#9767)
Description: this PR adds an option to select different process loggers based on bazel config. Notably it introduces an Android Logger that is able to log to Android App's log stream.
Risk Level: low. The new functionality is bazel config guarded.
Testing: local. CI
Docs Changes: documented in bazel's README.md
