+++ bazel/BUILD
@@ -323,6 +323,11 @@ config_setting(
     values = {"define": "manual_stamp=manual_stamp"},
 )
 
+config_setting(
+    name = "android_logger",
+    values = {"define": "logger=android"},
+)
+
 alias(
     name = "apple",
     actual = select(
