+++ source/common/common/BUILD
@@ -133,10 +133,41 @@ envoy_cc_library(
     hdrs = ["logger.h"],
     external_deps = ["abseil_synchronization"],
     deps = [
+        ":base_logger_lib",
         ":lock_guard_lib",
         ":macros",
         ":non_copyable",
-    ],
+    ] + select({
+        "//bazel:android_logger": ["logger_impl_lib_android"],
+        "//conditions:default": ["logger_impl_lib_standard"],
+    }),
+)
+
+envoy_cc_library(
+    name = "base_logger_lib",
+    srcs = ["base_logger.cc"],
+    hdrs = ["base_logger.h"],
+)
+
+envoy_cc_library(
+    name = "logger_impl_lib_standard",
+    hdrs = ["standard/logger_impl.h"],
+    strip_include_prefix = "standard",
+    deps = [":base_logger_lib"],
+)
+
+envoy_cc_library(
+    name = "logger_impl_lib_android",
+    srcs = select({
+        "//bazel:android_logger": ["android/logger_impl.cc"],
+        "//conditions:default": [],
+    }),
+    hdrs = select({
+        "//bazel:android_logger": ["android/logger_impl.h"],
+        "//conditions:default": [],
+    }),
+    strip_include_prefix = "android",
+    deps = [":base_logger_lib"],
 )
 
 envoy_cc_library(
