+++ source/common/common/standard/logger_impl.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "common/common/base_logger.h"
+
+namespace Envoy {
+namespace Logger {
+
+#define GENERATE_LOGGER(X) StandardLogger(#X),
+
+} // namespace Logger
+} // namespace Envoy
